requirements.txt
selenium==4.10.0
dotenv-python
webdriver_manager